chore: Refactor client models folder structure
