+++ client/src/app/routes/stardust/NFTDetails.tsx
@@ -3,7 +3,7 @@ import React, { ReactNode } from "react";
 import { RouteComponentProps } from "react-router-dom";
 import { ServiceFactory } from "../../../factories/serviceFactory";
 import { ClipboardHelper } from "../../../helpers/clipboardHelper";
-import { INftActivityHistory } from "../../../models/api/stardust/INftDetailsResponse";
+import { INftActivityHistory } from "../../../models/api/stardust/nft/INftDetailsResponse";
 import { STARDUST } from "../../../models/config/protocolVersion";
 import { StardustTangleCacheService } from "../../../services/stardust/stardustTangleCacheService";
 import AsyncComponent from "../../components/AsyncComponent";
