+++ client/src/models/api/stardust/foundry/IFoundryRequest.ts
