+++ client/src/models/api/stardust/nft/INftAddressDetailsRequest.ts
