+++ client/src/models/api/stardust/nft/INftDetailsRequest.ts
