+++ client/src/models/api/stardust/nft/INftDetailsResponse.ts
@@ -1,4 +1,4 @@
-import { IResponse } from "../IResponse";
+import { IResponse } from "../../IResponse";
 
 export interface INftActivityHistory {
     /**
