+++ client/src/models/api/stardust/nft/INftOutputsRequest.ts
