+++ client/src/services/stardust/stardustApiClient.ts
@@ -1,6 +1,10 @@
 import { FetchHelper } from "../../helpers/fetchHelper";
 import { IMilestoneDetailsRequest } from "../../models/api/IMilestoneDetailsRequest";
 import { IOutputDetailsRequest } from "../../models/api/IOutputDetailsRequest";
+import { IFoundriesRequest } from "../../models/api/stardust/foundry/IFoundriesRequest";
+import { IFoundriesResponse } from "../../models/api/stardust/foundry/IFoundriesResponse";
+import { IFoundryRequest } from "../../models/api/stardust/foundry/IFoundryRequest";
+import { IFoundryResponse } from "../../models/api/stardust/foundry/IFoundryResponse";
 import { IAliasRequest } from "../../models/api/stardust/IAliasRequest";
 import { IAliasResponse } from "../../models/api/stardust/IAliasResponse";
 import { IAssociatedOutputsRequest } from "../../models/api/stardust/IAssociatedOutputsRequest";
@@ -9,17 +13,7 @@ import { IBaseTokenGetRequest } from "../../models/api/stardust/IBaseTokenGetReq
 import { IBaseTokenGetResponse } from "../../models/api/stardust/IBaseTokenGetResponse";
 import { IBlockDetailsRequest } from "../../models/api/stardust/IBlockDetailsRequest";
 import { IBlockDetailsResponse } from "../../models/api/stardust/IBlockDetailsResponse";
-import { IFoundriesRequest } from "../../models/api/stardust/IFoundriesRequest";
-import { IFoundriesResponse } from "../../models/api/stardust/IFoundriesResponse";
-import { IFoundryRequest } from "../../models/api/stardust/IFoundryRequest";
-import { IFoundryResponse } from "../../models/api/stardust/IFoundryResponse";
 import { IMilestoneDetailsResponse } from "../../models/api/stardust/IMilestoneDetailsResponse";
-import { INftAddressDetailsRequest } from "../../models/api/stardust/INftAddressDetailsRequest";
-import { INftAddressDetailsResponse } from "../../models/api/stardust/INftAddressDetailsResponse";
-import { INftDetailsRequest } from "../../models/api/stardust/INftDetailsRequest";
-import { INftDetailsResponse } from "../../models/api/stardust/INftDetailsResponse";
-import { INftOutputsRequest } from "../../models/api/stardust/INftOutputsRequest";
-import { INftOutputsResponse } from "../../models/api/stardust/INftOutputsResponse";
 import { IOutputDetailsResponse } from "../../models/api/stardust/IOutputDetailsResponse";
 import { ISearchRequest } from "../../models/api/stardust/ISearchRequest";
 import { ISearchResponse } from "../../models/api/stardust/ISearchResponse";
@@ -27,6 +21,12 @@ import { ITransactionDetailsRequest } from "../../models/api/stardust/ITransacti
 import { ITransactionDetailsResponse } from "../../models/api/stardust/ITransactionDetailsResponse";
 import { ITransactionHistoryRequest } from "../../models/api/stardust/ITransactionHistoryRequest";
 import { ITransactionHistoryResponse } from "../../models/api/stardust/ITransactionHistoryResponse";
+import { INftAddressDetailsRequest } from "../../models/api/stardust/nft/INftAddressDetailsRequest";
+import { INftAddressDetailsResponse } from "../../models/api/stardust/nft/INftAddressDetailsResponse";
+import { INftDetailsRequest } from "../../models/api/stardust/nft/INftDetailsRequest";
+import { INftDetailsResponse } from "../../models/api/stardust/nft/INftDetailsResponse";
+import { INftOutputsRequest } from "../../models/api/stardust/nft/INftOutputsRequest";
+import { INftOutputsResponse } from "../../models/api/stardust/nft/INftOutputsResponse";
 import { IStatsGetRequest } from "../../models/api/stats/IStatsGetRequest";
 import { IStatsGetResponse } from "../../models/api/stats/IStatsGetResponse";
 import { ApiClient } from "../apiClient";
