+++ client/src/services/stardust/stardustTangleCacheService.ts
@@ -2,23 +2,23 @@
 import { IBlock, IBlockMetadata, IOutputResponse } from "@iota/iota.js-stardust";
 import { ServiceFactory } from "../../factories/serviceFactory";
 import { IBech32AddressDetails } from "../../models/api/IBech32AddressDetails";
+import { IFoundriesRequest } from "../../models/api/stardust/foundry/IFoundriesRequest";
+import { IFoundriesResponse } from "../../models/api/stardust/foundry/IFoundriesResponse";
+import { IFoundryRequest } from "../../models/api/stardust/foundry/IFoundryRequest";
+import { IFoundryResponse } from "../../models/api/stardust/foundry/IFoundryResponse";
 import { IAliasRequest } from "../../models/api/stardust/IAliasRequest";
 import { IAliasResponse } from "../../models/api/stardust/IAliasResponse";
 import { IAssociatedOutputsResponse } from "../../models/api/stardust/IAssociatedOutputsResponse";
-import { IFoundriesRequest } from "../../models/api/stardust/IFoundriesRequest";
-import { IFoundriesResponse } from "../../models/api/stardust/IFoundriesResponse";
-import { IFoundryRequest } from "../../models/api/stardust/IFoundryRequest";
-import { IFoundryResponse } from "../../models/api/stardust/IFoundryResponse";
 import { IMilestoneDetailsResponse } from "../../models/api/stardust/IMilestoneDetailsResponse";
-import { INftAddressDetailsRequest } from "../../models/api/stardust/INftAddressDetailsRequest";
-import { INftAddressDetailsResponse } from "../../models/api/stardust/INftAddressDetailsResponse";
-import { INftDetailsRequest } from "../../models/api/stardust/INftDetailsRequest";
-import { INftDetailsResponse } from "../../models/api/stardust/INftDetailsResponse";
-import { INftOutputsRequest } from "../../models/api/stardust/INftOutputsRequest";
-import { INftOutputsResponse } from "../../models/api/stardust/INftOutputsResponse";
 import { ISearchResponse } from "../../models/api/stardust/ISearchResponse";
 import { ITransactionHistoryRequest } from "../../models/api/stardust/ITransactionHistoryRequest";
 import { ITransactionHistoryResponse } from "../../models/api/stardust/ITransactionHistoryResponse";
+import { INftAddressDetailsRequest } from "../../models/api/stardust/nft/INftAddressDetailsRequest";
+import { INftAddressDetailsResponse } from "../../models/api/stardust/nft/INftAddressDetailsResponse";
+import { INftDetailsRequest } from "../../models/api/stardust/nft/INftDetailsRequest";
+import { INftDetailsResponse } from "../../models/api/stardust/nft/INftDetailsResponse";
+import { INftOutputsRequest } from "../../models/api/stardust/nft/INftOutputsRequest";
+import { INftOutputsResponse } from "../../models/api/stardust/nft/INftOutputsResponse";
 import { STARDUST } from "../../models/config/protocolVersion";
 import { TangleCacheService } from "../tangleCacheService";
 import { StardustApiClient } from "./stardustApiClient";
